Ale recommendations for doc generation
packloadall only changes when packages are loaded, during .vimrc load instead of after. This means that packages are able to be loaded so that 'heltags ALL' can pick up the docs to generate the help files. Packages are only loaded once anyway so it shouldn't be a big problem to change the order